[GPP] [FIX] Fix enemy turret detection

+++ Source/DiamondProject/Public/Enemy.h
@@ -27,7 +27,7 @@ protected:
     float AttackCooldown = 3;
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Enemy Attack")
-    float AttackDamage = 1;
+    int AttackDamage = 1;
 
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "Enemy Behavior")
 	float BaseDetectionRange = 300;
